+++ di.py
@@ -1,5 +1,6 @@
 from dependency_injector import containers, providers
 from data.purpleair import PurpleAirDataResolver
+from data.envcanada import EnvironmentCanadaDataResolver
 from pixelperfectpi import Clock
 
 class Container(containers.DeclarativeContainer):
@@ -10,7 +11,10 @@ class Container(containers.DeclarativeContainer):
         url=config.purpleair.url,
     )
 
+    env_canada = providers.Singleton(EnvironmentCanadaDataResolver)
+
     clock = providers.Singleton(
         Clock,
         purpleair=purpleair,
+        env_canada=env_canada,
     )
